+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 #### 1.1.0 (UNRELEASED)
 
+* Fixed `write` to cap SVG width, height and viewBox to a minimum of 1px 
 * Fixed grouping of `stat` command in `vpype --help`
+* Bump svgelements from 1.3.2 to 1.3.4 (#101)
 
 
 #### 1.0.0 (2020-11-29)